requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.7
Django==5.0.6
django-ninja==1.1.0
django-ninja-extra==0.20.7
django-ninja-jwt==5.3.1
dnspython==2.6.1
email_validator==2.1.1
idna==3.7
injector==0.21.0
libsql-experimental==0.0.34
ninja-schema==0.13.6
psycopg-binary==3.1.19
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.0
